p_drop fixes, extending source_break scenarios

+++ Emulation Scripts/source_break-acs-01.sh
@@ -13,18 +13,21 @@ BATCH_SIZE=16
 
 for PATHS in 1 2
 do
-    for CONNECTIONS in 1 2
+    for CONNECTIONS in 1 2 4
     do
-        if [ "$CONNECTIONS" -lt "$(($SOURCES * $PATHS))" ]
+        if [ "$CONNECTIONS" -lt "$(($SOURCES * $PATHS))"]
         then
-            path="../evaluation/data/${DIRECTORY}/s${SOURCES}_p${PATHS}_c${CONNECTIONS}_b${BATCH_SIZE}_p${P_DROP}/"
-            echo $path
-            rm -r $path
-            mkdir -p $path
-            for i in {0..9}
-            do
-                java -Djava.util.logging.SimpleFormatter.format='%5$s%6$s%n' MSTCP.vegas.more.Experiment $DIRECTORY $SOURCES $PATHS $CONNECTIONS $BATCH_SIZE $P_DROP $PACKET_LIMIT $i $FILE
-            done
+            if [ "$PATHS" -le "$CONNECTIONS"]
+            then
+                path="../evaluation/data/${DIRECTORY}/s${SOURCES}_p${PATHS}_c${CONNECTIONS}_b${BATCH_SIZE}_p${P_DROP}/"
+                echo $path
+                rm -r $path
+                mkdir -p $path
+                for i in {0..99}
+                do
+                    java -Djava.util.logging.SimpleFormatter.format='%5$s%6$s%n' MSTCP.vegas.more.Experiment $DIRECTORY $SOURCES $PATHS $CONNECTIONS $BATCH_SIZE $P_DROP $PACKET_LIMIT $i $FILE
+                done
+            fi
         fi
     done
 done
